fix(frontend): fix bug with blocker bb-473
