+++ apps/frontend/src/pages/profile/libs/components/update-password-form/update-password-form.tsx
@@ -1,9 +1,8 @@
-import runImg from "~/assets/img/run.svg";
-import { Button, Input, Popup } from "~/libs/components/components.js";
+import { Button, Input } from "~/libs/components/components.js";
 import {
 	useAppForm,
-	useBlocker,
 	useCallback,
+	useEffect,
 	useState,
 } from "~/libs/hooks/hooks.js";
 import {
@@ -18,9 +17,13 @@ import styles from "./styles.module.css";
 
 type Properties = {
 	onSubmit: (payload: UserUpdatePasswordRequestDto) => void;
+	setIsDirty: (payload: boolean) => void;
 };
 
-const UpdatePasswordForm: React.FC<Properties> = ({ onSubmit }: Properties) => {
+const UpdatePasswordForm: React.FC<Properties> = ({
+	onSubmit,
+	setIsDirty,
+}: Properties) => {
 	const [isCurrentPasswordVisible, setIsCurrentPasswordVisible] =
 		useState<boolean>(false);
 	const [isNewPasswordVisible, setIsNewPasswordVisible] =
@@ -28,7 +31,7 @@ const UpdatePasswordForm: React.FC<Properties> = ({ onSubmit }: Properties) => {
 	const [isConfirmNewPasswordVisible, setIsConfirmNewPasswordVisible] =
 		useState<boolean>(false);
 
-	const { control, errors, getValues, handleSubmit, isDirty, reset, setError } =
+	const { control, errors, getValues, handleSubmit, isDirty, setError } =
 		useAppForm<UserUpdatePasswordFormDto>({
 			defaultValues: DEFAULT_UPDATE_PASSWORD_PAYLOAD,
 			validationSchema: userUpdatePasswordValidationSchema,
@@ -65,68 +68,44 @@ const UpdatePasswordForm: React.FC<Properties> = ({ onSubmit }: Properties) => {
 		setIsConfirmNewPasswordVisible((previousState) => !previousState);
 	}, []);
 
-	const blocker = useBlocker(isDirty);
-
-	const handleCancelPopupClick = useCallback((): void => {
-		if (blocker.state === "blocked") {
-			blocker.reset();
-		}
-	}, [blocker]);
-
-	const handleConfirmPopupClick = useCallback((): void => {
-		reset();
-
-		if (blocker.state === "blocked") {
-			blocker.proceed();
-		}
-	}, [blocker, reset]);
+	useEffect(() => {
+		setIsDirty(isDirty);
+	}, [isDirty, setIsDirty]);
 
 	return (
-		<>
-			<Popup
-				closeButtonLabel="CANCEL"
-				confirmButtonLabel="YES"
-				hasCloseIcon
-				icon={runImg}
-				isOpen={isDirty && blocker.state === "blocked"}
-				onClose={handleCancelPopupClick}
-				onConfirm={handleConfirmPopupClick}
-				title="Unsaved changes will be lost. Continue?"
+		<form className={styles["form"]} onSubmit={handleFormSubmit}>
+			<Input
+				control={control}
+				errors={errors}
+				iconName={isCurrentPasswordVisible ? "crossedEye" : "eye"}
+				label="Current Password"
+				name="currentPassword"
+				onIconClick={handleTogglePasswordVisibility}
+				placeholder="••••••"
+				type={isCurrentPasswordVisible ? "text" : "password"}
 			/>
-			<form className={styles["form"]} onSubmit={handleFormSubmit}>
-				<Input
-					control={control}
-					errors={errors}
-					iconName={isCurrentPasswordVisible ? "crossedEye" : "eye"}
-					label="Current Password"
-					name="currentPassword"
-					onIconClick={handleTogglePasswordVisibility}
-					placeholder="••••••"
-					type={isCurrentPasswordVisible ? "text" : "password"}
-				/>
-				<Input
-					control={control}
-					errors={errors}
-					iconName={isNewPasswordVisible ? "crossedEye" : "eye"}
-					label="New Password"
-					name="newPassword"
-					onIconClick={handleToggleNewPasswordVisibility}
-					placeholder="••••••"
-					type={isNewPasswordVisible ? "text" : "password"}
-				/>
-				<Input
-					control={control}
-					errors={errors}
-					iconName={isConfirmNewPasswordVisible ? "crossedEye" : "eye"}
-					label="Confirm New Password"
-					name="confirmNewPassword"
-					onIconClick={handleToggleConfirmNewPasswordVisibility}
-					placeholder="••••••"
-					type={isConfirmNewPasswordVisible ? "text" : "password"}
-				/>
-				<Button label="SAVE" type="submit" />
-			</form>
-		</>
+			<Input
+				control={control}
+				errors={errors}
+				iconName={isNewPasswordVisible ? "crossedEye" : "eye"}
+				label="New Password"
+				name="newPassword"
+				onIconClick={handleToggleNewPasswordVisibility}
+				placeholder="••••••"
+				type={isNewPasswordVisible ? "text" : "password"}
+			/>
+			<Input
+				control={control}
+				errors={errors}
+				iconName={isConfirmNewPasswordVisible ? "crossedEye" : "eye"}
+				label="Confirm New Password"
+				name="confirmNewPassword"
+				onIconClick={handleToggleConfirmNewPasswordVisibility}
+				placeholder="••••••"
+				type={isConfirmNewPasswordVisible ? "text" : "password"}
+			/>
+			<Button label="SAVE" type="submit" />
+		</form>
 	);
 };
