+++ apps/frontend/src/pages/profile/libs/components/update-user-form/update-user-form.tsx
@@ -1,6 +1,5 @@
-import runImg from "~/assets/img/run.svg";
-import { Button, Input, Popup } from "~/libs/components/components.js";
-import { useAppForm, useBlocker, useCallback } from "~/libs/hooks/hooks.js";
+import { Button, Input } from "~/libs/components/components.js";
+import { useAppForm, useCallback, useEffect } from "~/libs/hooks/hooks.js";
 import {
 	type UserDto,
 	type UserUpdateFormDto,
@@ -12,15 +11,17 @@ import styles from "./styles.module.css";
 
 type Properties = {
 	onSubmit: (payload: UserUpdateRequestDto) => void;
+	setIsDirty: (payload: boolean) => void;
 	user: UserDto;
 };
 
 const UpdateUserForm: React.FC<Properties> = ({
 	onSubmit,
+	setIsDirty,
 	user,
 }: Properties) => {
 	const { email, name } = user;
-	const { control, errors, handleSubmit, isDirty, reset } =
+	const { control, errors, handleSubmit, isDirty } =
 		useAppForm<UserUpdateFormDto>({
 			defaultValues: { email, name },
 			validationSchema: userUpdateValidationSchema,
@@ -33,55 +34,31 @@ const UpdateUserForm: React.FC<Properties> = ({
 		[handleSubmit, onSubmit],
 	);
 
-	const blocker = useBlocker(isDirty);
-
-	const handleCancelPopupClick = useCallback((): void => {
-		if (blocker.state === "blocked") {
-			blocker.reset();
-		}
-	}, [blocker]);
-
-	const handleConfirmPopupClick = useCallback((): void => {
-		reset();
-
-		if (blocker.state === "blocked") {
-			blocker.proceed();
-		}
-	}, [blocker, reset]);
+	useEffect(() => {
+		setIsDirty(isDirty);
+	}, [isDirty, setIsDirty]);
 
 	return (
-		<>
-			<Popup
-				closeButtonLabel="CANCEL"
-				confirmButtonLabel="YES"
-				hasCloseIcon
-				icon={runImg}
-				isOpen={isDirty && blocker.state === "blocked"}
-				onClose={handleCancelPopupClick}
-				onConfirm={handleConfirmPopupClick}
-				title="Unsaved changes will be lost. Continue?"
+		<form className={styles["form"]} onSubmit={handleFormSubmit}>
+			<Input
+				control={control}
+				errors={errors}
+				label="Name"
+				name="name"
+				placeholder="name"
+				type="text"
 			/>
-			<form className={styles["form"]} onSubmit={handleFormSubmit}>
-				<Input
-					control={control}
-					errors={errors}
-					label="Name"
-					name="name"
-					placeholder="name"
-					type="text"
-				/>
-				<Input
-					control={control}
-					errors={errors}
-					isDisabled
-					label="Email"
-					name="email"
-					placeholder="email"
-					type="email"
-				/>
-				<Button label="SAVE" type="submit" />
-			</form>
-		</>
+			<Input
+				control={control}
+				errors={errors}
+				isDisabled
+				label="Email"
+				name="email"
+				placeholder="email"
+				type="email"
+			/>
+			<Button label="SAVE" type="submit" />
+		</form>
 	);
 };
